requirements.txt
aioeasywebdav==2.4.0
python-dotenv==0.7.1
paho-mqtt==1.3.1
requests==2.18.4